+++ src/locale/ug/index.js
@@ -9,6 +9,7 @@ import match from './_lib/match/index.js'
  * @category Locales
  * @summary Uighur locale
  * @language Uighur
+ * @iso-639-2 uig
  * @author Abduwaly M. [@abduwaly]{@link https://github.com/abduwaly}
  */
 var locale = {