+++ src/locale/ug/test.js
@@ -4,4 +4,4 @@
 // import assert from 'power-assert'
 // import locale from '.'
 
-describe.skip('ui locale', function() {})
+describe.skip('ug locale', function() {})